Added statement to manually configure the MAC address for IRB interfaces with virtual-gateways i.e. EVPN Anycast GWs configured.
This is to change behaviour whereby the switch will send frames, including
ARP responses for the Anycast GW address itself, from the burned in
interface address.  While this is not an issue for us, with L3 access
layer and servers connected directly to device with IRB, it causes problems
if there is a switch device between the IRB interface and end host.  As
that switch will never see the IRB VGA MAC, and thus will have to flood
unicast frames sent to the gateway by end servers with that as dest MAC.

As mentioned this should not be an issue in our setup, but it seems like
a sensible best practice and might avoid some weird edge cases, perhaps
with say Ganeti hosts with on-device bridges configured.

https://www.juniper.net/documentation/us/en/software/junos/evpn-vxlan/topics/ref/statement/virtual-gateway-v4-mac-edit-interfaces.html